chore: moved Default Scene & Platforms rendering out from the BuildConfigurationWindow to separate views
