+++ com.stansassets.scene-management/BuildConfigurator/Editor/Views.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 06efb691de634746818ca9b840906313
+timeCreated: 1688733392
