+++ com.stansassets.scene-management/BuildConfigurator/Editor/Views/DefaultScenesView.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 30514be724d54aab8d986670b36b4266
+timeCreated: 1688728304
