+++ com.stansassets.scene-management/BuildConfigurator/Editor/Views/PlatformsView.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 2efff43a5a68493190f5897cac9e24ba
+timeCreated: 1688732085
