+++ com.stansassets.scene-management/BuildConfigurator/Editor/Views/SettingsView.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: caaa00cccaf3495594cc2ffb6b7f24c3
+timeCreated: 1688733377
